Rewrite documentation (#227)
* Rewrote documentation

Added a guide to contributing to this project, as well as documentation on the abbreviation file format. Moved the grammar file documentation to the grammar folder. Rewrote large sections of the readme, relegating some content to other Markdown documents, and renamed the file for consistency.

* Fixed code blocks in grammar documentation

The code blocks were causing colored diffs on GitHub to go haywire.

* Simplified language

Also added next steps after submitting a localization.

* Added examples to abbreviation documentation

* Fixed link